+++ src/ClusterApi.php
@@ -44,11 +44,6 @@ class ClusterApi
         return new Endpoints\Cmses($this->client);
     }
 
-    public function commands(): Endpoints\Commands
-    {
-        return new Endpoints\Commands($this->client);
-    }
-
     public function crons(): Endpoints\Crons
     {
         return new Endpoints\Crons($this->client);