+++ src/Models/Cluster.php
@@ -17,7 +17,6 @@ class Cluster extends ClusterModel implements Model
     private ?string $customerId = null;
     private ?bool $wordpressToolkitEnabled = null;
     private ?bool $databaseToolkitEnabled = null;
-    private ?bool $commandToolkitEnabled = null;
     private ?bool $malwareToolkitEnabled = null;
     private ?bool $malwareToolkitScansEnabled = null;
     private ?bool $bubblewrapToolkitEnabled = null;
@@ -145,18 +144,6 @@ class Cluster extends ClusterModel implements Model
         return $this;
     }
 
-    public function isCommandToolkitEnabled(): ?bool
-    {
-        return $this->commandToolkitEnabled;
-    }
-
-    public function setCommandToolkitEnabled(?bool $commandToolkitEnabled): Cluster
-    {
-        $this->commandToolkitEnabled = $commandToolkitEnabled;
-
-        return $this;
-    }
-
     public function istMalwareToolkitEnabled(): ?bool
     {
         return $this->malwareToolkitEnabled;
@@ -242,7 +229,6 @@ class Cluster extends ClusterModel implements Model
             ->setCustomerId(Arr::get($data, 'customer_id'))
             ->setWordpressToolkitEnabled(Arr::get($data, 'wordpress_toolkit_enabled'))
             ->setDatabaseToolkitEnabled(Arr::get($data, 'database_toolkit_enabled'))
-            ->setCommandToolkitEnabled(Arr::get($data, 'command_toolkit_enabled'))
             ->setMalwareToolkitEnabled(Arr::get($data, 'malware_toolkit_enabled'))
             ->setMalwareToolkitScansEnabled(Arr::get($data, 'malware_toolkit_scans_enabled'))
             ->setBubblewrapToolkitEnabled(Arr::get($data, 'bubblewrap_toolkit_enabled'))
@@ -264,7 +250,6 @@ class Cluster extends ClusterModel implements Model
             'customer_id' => $this->getCustomerId(),
             'wordpress_toolkit_enabled' => $this->isWordpressToolkitEnabled(),
             'database_toolkit_enabled' => $this->isDatabaseToolkitEnabled(),
-            'command_toolkit_enabled' => $this->isCommandToolkitEnabled(),
             'malware_toolkit_enabled' => $this->istMalwareToolkitEnabled(),
             'malware_toolkit_scans_enabled' => $this->isMalwareToolkitScansEnabled(),
             'bubblewrap_toolkit_enabled ' => $this->isBubblewrapToolkitEnabled(),